backend/requirements.txt
Pillow==9.4.0
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-phonenumber-field==7.0.2
phonenumberslite==8.13.7
django-filter==22.1
django-cors-headers==3.14.0
